requirements.txt
customtkinter
pystray
requests
Pillow
pynput